+++ MISO/Core/OUDS/Components/Sources/Navigations/ListItem/Nested/MISOListItemFlag.swift
@@ -64,15 +64,10 @@ public struct MISOListItemFlag: View {
     // MARK: Body
 
     public var body: some View {
-        HStack {
-            MISOImage(asset: asset, accessibilityLabel: description, renderingMode: .original)
-                .aspectRatio(contentMode: .fit)
-                .frame(height: flagHeight)
-                .opacity(opacity)
-        }
-        .accessibilityElement(children: .ignore)
-        .accessibilityLabel(description)
-        .frame(minWidth: minWidth, alignment: .center)
+        MISOImage(asset: asset, accessibilityLabel: description, renderingMode: .original)
+            .aspectRatio(contentMode: .fit)
+            .frame(width: flagWidth)
+            .opacity(opacity)
     }
 
     // MARK: Helpers
@@ -81,11 +76,7 @@ public struct MISOListItemFlag: View {
         isEnabled ? theme.opacities.opaque : theme.opacities.disabled
     }
 
-    private var flagHeight: SizeSemanticToken {
-        theme.listItem.sizeFlagHeight * dynamicTypeSize.percentageRate / 100
-    }
-
-    private var minWidth: SizeSemanticToken {
+    private var flagWidth: SizeSemanticToken {
         let rawSize = if itemSize == .small {
             theme.listItem.sizeAssetSmall
         } else {