+++ MISO/Core/OUDS/Components/Sources/Navigations/ListItem/Nested/MISOListItemIcon.swift
@@ -232,12 +232,12 @@ public struct MISOListItemIcon: View {
         case .default:
             switch size {
             case .medium:
-                theme.listItem.sizeAssetMedium
+                theme.sizes.iconWithLabelLargeSizeMedium
             case .large:
-                theme.listItem.sizeAssetLarge
+                theme.sizes.iconWithLabelLargeSizeLarge
             }
         case .small:
-            theme.listItem.sizeAssetSmall
+            theme.sizes.iconWithLabelLargeSizeSmall
         }
 
         return rawSize * dynamicTypeSize.percentageRate / 100