+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0](https://github.com/pylapp/miso-ios/releases/comapre/1.0.0..1.1.0) - 2026-09-22
+## [Unreleased](hhttps://github.com/pylapp/miso-ios/releases/compare/1.1.0...develop)
+
+### Changed
+
+- `list item` component to version v1.2.0 (Orange-OpenSource/ouds-ios#1714)
+
+### Fixed
+
+- **BREAKING**: missing `maxWidthBoxedText` token in `SizeMultipleSemanticTokens` (Orange-OpenSource/ouds-ios#1714)
+
+## [1.1.0](https://github.com/pylapp/miso-ios/releases/tag/1.0.0) - 2026-09-22
 
 ### Added